From Albert Bursche
there is a bug in TTree::GetMaximum() line 5205

Double_t cmax = -FLT_MAX;

http://root.cern.ch/root/html/src/TTree.cxx.html#PNz1z

I should be initialised with -std::numeric_limits<double>::infinity() or
at least with  -std::numeric_limits<double>::max() or -DBL_MAX for that
matter. In the current implementation the function returns the wrong
value if the maximum is less then -FLT_MAX; which is possible even for
finite numbers in double precision.


git-svn-id: http://root.cern.ch/svn/root/trunk@47980 27541ba8-7e3a-0410-8455-c3a389f83636

+++ tree/tree/src/TTree.cxx
@@ -5241,7 +5241,7 @@ Double_t TTree::GetMaximum(const char* columname)
       return 0;
    }
    TBranch* branch = leaf->GetBranch();
-   Double_t cmax = -FLT_MAX;
+   Double_t cmax = -DBL_MAX;
    for (Long64_t i = 0; i < fEntries; ++i) {
       Long64_t entryNumber = this->GetEntryNumber(i);
       if (entryNumber < 0) break;
